pyproject.toml
[tool.poetry]
name = "immigration-game"
version = "0.1.0"
description = "Cosmic Bureaucracy: Lost in Translation"
authors = ["D. Lowl <d.lowl@d-lowl.space>"]
readme = "README.md"
packages = [{include = "cblit"}]

[tool.poetry.scripts]
cblit = "cblit.main:app"

[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = "^0.7.0"}
openai = "^0.27.2"
typing-extensions = "^4.5.0"
dataclasses-json = "^0.5.7"
mypy = "^1.1.1"
loguru = "^0.6.0"
dacite = "^1.8.0"
python-socketio = "^5.8.0"
sanic = "^23.3.0"
langchain = "^0.0.188"
faiss-cpu = "^1.7.4"
pytest-asyncio = "^0.21.0"
retry = "^0.9.2"
types-retry = "^0.9.9.3"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.0"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
show_error_codes = true
exclude = ["test"]
strict = true
disallow_untyped_decorators = false  # ignore functions using decorators
ignore_missing_imports = true  # silence all import errors related to untyped libraries

[tool.ruff]
line-length = 120
select = [
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
    "Q",    # flake8-quotes
    "ISC",  # flake8-implicit-str-concat
    "F",    # pyflakes
    "D",    # pydocstyle
    "E",    # pycodestyle error
    "W",    # pycodestyle warning
    "N",    # pep8-naming
    "I",    # isort
    "PL",   # pylint rules from categories "Convention", "Error", and "Warning"
    "PLE",  # ruff currently implements only a subset of pylint's rules
    "PLW",  # pylint warning
    "PLR",  # pylint refactor
    "UP",   # pyupgrade
    "C",    # Complexity (mccabe+) & comprehensions
]

[tool.ruff.pydocstyle]
convention = "google"
